fix: resolve checks failing on the merged tree
Merging origin/master surfaced check failures, all pre-existing on master
(verified: pure origin/master fails the same typecheck+lint) plus one knip
item from this branch. Faithful fixes (no lint gaming):

- knip: ignore session-order.ts exports (consumed only by SessionsList.svelte,
  which knip can't parse) — this branch's only own failure
- tests/bot.test.ts: import afterEach (used but missing from the bun:test
  import → TS2304 + oxlint typescript errors)
- src/chat/types.ts (304 > 300 lines): extract AuthorizationDenyReason /
  AuthorizationResult into src/chat/authorization-types.ts and re-export
- src/auth.ts: extract openDmAutoAdd helper so checkAuthorizationExtended is
  back under the 50-line limit
- tests/plugins/discovery.test.ts: skip the chmod(0) unreadable-file case when
  running as root (root bypasses permission bits; the assertion can't hold)
